requirements.txt
Flask==2.3.3
Dash==2.13.1
python-dotenv==1.0.0
plotly==5.15.0
requests==2.31.0
pandas==2.1.2